requirements.txt
Flask==3.0.0
pysrt==1.1.2
deep-translator==1.11.4
gTTS==2.5.1
google-generativeai==0.8.3
edge-tts==6.1.12
